refactor: move SierpinskiFillProvider implementation to new cpp file
