+++ CMakeLists.txt
@@ -145,6 +145,7 @@ set(engine_SRCS # Except main.cpp.
     src/infill/NoZigZagConnectorProcessor.cpp
     src/infill/ZigzagConnectorProcessor.cpp
     src/infill/SierpinskiFill.cpp
+    src/infill/SierpinskiFillProvider.cpp
     src/infill/SpaghettiInfill.cpp
     src/infill/SpaghettiInfillPathGenerator.cpp
     src/infill/SubDivCube.cpp
