+++ src/infill/SierpinskiFillProvider.h
@@ -8,6 +8,7 @@
 #include "DensityProvider.h"
 #include "ImageBasedDensityProvider.h"
 #include "UniformDensityProvider.h"
+#include "../settings/settings.h"
 
 namespace cura
 {
@@ -23,87 +24,31 @@ namespace cura
 class SierpinskiFillProvider
 {
     static constexpr bool get_constructor = true;
-    static constexpr bool use_dithering = true;
+    static constexpr bool use_dithering = true; // !< Whether to employ dithering and error propagation
 protected:
+    //! Basic parameters from which to start constructing the sierpinski fractal
     struct FractalConfig
     {
-        int depth;
-        AABB aabb; 
+        int depth; //!< max recursion depth
+        AABB aabb; //!< The bounding box of the initial Triangles in the Sierpinski curve
     };
 public:
     FractalConfig fractal_config;
-    DensityProvider* density_provider;
-    std::optional<SierpinskiFill> fill_pattern_for_all_layers;
+    DensityProvider* density_provider; //!< The object which determines the requested density at each region
+    std::optional<SierpinskiFill> fill_pattern_for_all_layers; //!< The fill pattern if one and the same pattern is used on all layers
 
-    SierpinskiFillProvider(const AABB3D aabb_3d, coord_t min_line_distance, const coord_t line_width)
-    : fractal_config(getFractalConfig(aabb_3d, min_line_distance))
-    , density_provider(new UniformDensityProvider((float)line_width / min_line_distance))
-    , fill_pattern_for_all_layers(get_constructor, *density_provider, fractal_config.aabb, fractal_config.depth, line_width, use_dithering)
-    {
-    }
+    SierpinskiFillProvider(const AABB3D aabb_3d, coord_t min_line_distance, const coord_t line_width);
 
-    SierpinskiFillProvider(const AABB3D aabb_3d, coord_t min_line_distance, coord_t line_width, std::string cross_subdisivion_spec_image_file)
-    : fractal_config(getFractalConfig(aabb_3d, min_line_distance))
-    , density_provider(new ImageBasedDensityProvider(cross_subdisivion_spec_image_file, aabb_3d.getAABB()))
-    , fill_pattern_for_all_layers(get_constructor, *density_provider, fractal_config.aabb, fractal_config.depth, line_width, use_dithering)
-    {
-    }
+    SierpinskiFillProvider(const AABB3D aabb_3d, coord_t min_line_distance, coord_t line_width, std::string cross_subdisivion_spec_image_file);
 
-    Polygon generate(EFillMethod pattern, coord_t z, coord_t line_width) const
-    {
-        if (fill_pattern_for_all_layers)
-        {
-            if (pattern == EFillMethod::CROSS_3D)
-            {
-                return fill_pattern_for_all_layers->generateCross(z, line_width / 2);
-            }
-            else
-            {
-                return fill_pattern_for_all_layers->generateCross();
-            }
-        }
-        else
-        {
-            Polygon ret;
-            logError("Different density sierpinski fill for different layers is not implemented yet!\n");
-            std::exit(-1);
-            return ret;
-        }
-    }
+    Polygon generate(EFillMethod pattern, coord_t z, coord_t line_width) const;
 
-    ~SierpinskiFillProvider()
-    {
-        if (density_provider)
-        {
-            delete density_provider;
-        }
-    }
+    ~SierpinskiFillProvider();
 protected:
-    FractalConfig getFractalConfig(const AABB3D aabb_3d, coord_t min_line_distance)
-    {
-        AABB model_aabb = aabb_3d.getAABB();
-        Point model_aabb_size = model_aabb.max - model_aabb.min;
-        coord_t max_side_length = std::max(model_aabb_size.X, model_aabb_size.Y);
-        Point model_middle = model_aabb.getMiddle();
-
-        int depth = 0;
-        coord_t aabb_size = min_line_distance;
-        while (aabb_size < max_side_length)
-        {
-            aabb_size *= 2;
-            depth += 2;
-        }
-        const float sqrt2 = .5 * std::sqrt(2.0);
-        if (aabb_size * sqrt2 >= max_side_length)
-        {
-            aabb_size *= sqrt2;
-            depth--;
-        }
-
-        Point radius(aabb_size / 2, aabb_size / 2);
-        AABB aabb(model_middle - radius, model_middle + radius);
-        return FractalConfig{depth, aabb};
-    }
+    /*!
+     * Get the parameters with which to generate a sierpinski fractal for this object
+     */
+    FractalConfig getFractalConfig(const AABB3D aabb_3d, coord_t min_line_distance);
 };
 } // namespace cura
